# packages/contracts/openapi/api.v1.yaml
openapi: 3.1.0
info:
  title: NexusKB API
  version: 1.0.0
security:
  - bearerAuth: []
paths:
...
  /v1/access/users:
    get:
      summary: List observed authenticated users within the caller's tenant or department scope
      parameters:
        - { name: query, in: query, schema: { type: string, minLength: 1, maxLength: 128 } }
        - { name: department, in: query, schema: { type: string, minLength: 1, maxLength: 128 } }
        - { name: offset, in: query, schema: { type: integer, minimum: 0, default: 0 } }
        - {
            name: limit,
            in: query,
            schema: { type: integer, minimum: 1, maximum: 100, default: 25 },
          }
      responses:
        '200':
          description: Tenant-isolated or department-limited user directory page
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserDirectoryQueryResponse' }
        '400':
          description: Invalid user directory query
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /v1/access/users/{userId}/roles:
    patch:
      summary: Replace application-managed roles for an observed tenant user
      parameters:
        - { name: userId, in: path, required: true, schema: { type: string, maxLength: 256 } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserRoleUpdateRequest' }
      responses:
        '200':
          description: Managed roles updated with last-platform-admin protection
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserRoleUpdateResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { description: Last platform administrator cannot be removed }
  /v1/access/departments:
    get:
      summary: List tenant or department scoped effective department policies
      responses:
        '200':
          description: Department policy summaries
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DepartmentPolicyListResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /v1/access/departments/{department}:
    patch:
      summary: Restrict a department sensitivity policy
      parameters:
        - { name: department, in: path, required: true, schema: { type: string, maxLength: 128 } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DepartmentPolicyUpdateRequest' }
      responses:
        '200':
          description: Effective department policy
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DepartmentPolicyUpdateResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /v1/history/conversations:
    get:
      summary: List only the authenticated user's conversations
      parameters:
        - { name: query, in: query, schema: { type: string, minLength: 1, maxLength: 200 } }
        - { name: from, in: query, schema: { type: string, format: date-time } }
        - { name: to, in: query, schema: { type: string, format: date-time } }
        - { name: offset, in: query, schema: { type: integer, minimum: 0, default: 0 } }
        - {
            name: limit,
            in: query,
            schema: { type: integer, minimum: 1, maximum: 100, default: 20 },
          }
      responses:
        '200':
          description: User-owned conversation page
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ConversationListResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /v1/history/conversations/{conversationId}:
    parameters:
      - { name: conversationId, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      summary: Get a user-owned conversation
      responses:
        '200':
          description: Conversation and turns
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ConversationDetail' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      summary: Delete a user-owned conversation and its turns
      responses:
        '200':
          description: Conversation deleted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ConversationDeleteResponse' }
...
  /v1/documents:
    get:
      summary: List ACL-visible documents with server-side filtering and pagination
      parameters:
        - { name: search, in: query, schema: { type: string, maxLength: 200 } }
        - { name: department, in: query, schema: { type: string, maxLength: 128 } }
        - { name: sensitivity, in: query, schema: { $ref: '#/components/schemas/Sensitivity' } }
        - { name: status, in: query, schema: { $ref: '#/components/schemas/DocumentStatus' } }
        - { name: format, in: query, schema: { $ref: '#/components/schemas/DocumentFormat' } }
        - { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
        - {
            name: pageSize,
            in: query,
            schema: { type: integer, minimum: 1, maximum: 100, default: 20 },
          }
      responses:
        '200':
          description: ACL-filtered document page
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DocumentListResponse' }
        '400': { description: Invalid list query }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      summary: Upload a document and enqueue ingestion
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        '202':
          description: Queued
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UploadAccepted' }
        '409':
          description: Content already exists in the same tenant and ACL scope
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /v1/documents/upload-options:
    get:
      summary: Return server-owned upload limits and identity-scoped metadata options
      responses:
        '200':
          description: Upload options for the authenticated identity
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DocumentUploadOptions' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /v1/documents/{documentId}:
    parameters:
      - $ref: '#/components/parameters/DocumentId'
    get:
      summary: Get a tenant-scoped document
      responses:
        '200':
          description: Document metadata and version history
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DocumentDetail' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      summary: Idempotently delete a tenant-scoped document
      responses:
        '200':
          description: Deleted or already absent
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DocumentDeleteResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /v1/documents/{documentId}/reindex:
    parameters:
      - $ref: '#/components/parameters/DocumentId'
    post:
      summary: Create and enqueue a new version while the active version remains queryable
      responses:
        '202':
          description: Reindex queued
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReindexAccepted' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Document is not active or another ingestion job is running
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
  /v1/documents/{documentId}/metadata:
    parameters:
      - $ref: '#/components/parameters/DocumentId'
    patch:
      summary: Update authorized ACL metadata and enqueue a safe new document version
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [department, sensitivity]
              properties:
                department: { type: string, minLength: 1, maxLength: 128 }
                sensitivity: { $ref: '#/components/schemas/Sensitivity' }
      responses:
        '202':
          description: Metadata updated and reindex queued
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReindexAccepted' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /v1/ingestion-jobs:
    get:
      summary: List ACL-visible ingestion jobs with server-side pagination
      parameters:
        - { name: documentId, in: query, schema: { type: string, format: uuid } }
        - { name: status, in: query, schema: { $ref: '#/components/schemas/IngestionStatus' } }
        - { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
        - {
            name: pageSize,
            in: query,
            schema: { type: integer, minimum: 1, maximum: 100, default: 20 },
          }
      responses:
        '200':
          description: ACL-filtered ingestion job page
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IngestionJobListResponse' }
        '400': { description: Invalid ingestion job query }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /v1/ingestion-jobs/{jobId}:
    get:
      summary: Get a tenant-scoped ingestion job
      parameters:
        - name: jobId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Ingestion job
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IngestionJob' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/ingestion-jobs/{jobId}/retry:
    post:
      summary: Requeue an ACL-visible retryable failed ingestion job
      parameters:
        - name: jobId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '202':
          description: Retry queued
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IngestionRetryAccepted' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { description: Job is not failed or retryable }
  /v1/ingestion-jobs/failed:
    get:
      summary: List recent failed ingestion jobs in the authenticated tenant
      responses:
        '200':
          description: Up to 50 failed jobs ordered by most recent update
          content:
            application/json:
              schema:
                type: object
                required: [jobs]
                properties:
                  jobs:
                    type: array
                    items: { $ref: '#/components/schemas/IngestionJobSummary' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    DocumentId:
      name: documentId
      in: path
      required: true
      schema: { type: string, format: uuid }
  schemas:
    Sensitivity:
      type: string
      enum: [public, internal, confidential]
...
    UserDirectoryEntry:
      type: object
      additionalProperties: false
      required: [userId, department, roles, roleSource, status, lastAuthenticatedAt]
      properties:
        userId: { type: string, minLength: 1, maxLength: 256 }
        department: { type: string, minLength: 1, maxLength: 128 }
        roles:
          type: array
          maxItems: 32
          items: { type: string, minLength: 1, maxLength: 64 }
        roleSource: { type: string, enum: [identity, managed] }
        status: { const: observed }
        lastAuthenticatedAt: { type: string, format: date-time }
    UserDirectoryQueryResponse:
      type: object
      additionalProperties: false
      required: [users, total, offset, limit, scope]
      properties:
        users:
          type: array
          maxItems: 100
          items: { $ref: '#/components/schemas/UserDirectoryEntry' }
        total: { type: integer, minimum: 0 }
        offset: { type: integer, minimum: 0 }
        limit: { type: integer, minimum: 1, maximum: 100 }
        scope: { type: string, enum: [tenant, department] }
    UserRoleUpdateRequest:
      type: object
      additionalProperties: false
      required: [roles]
      properties:
        roles:
          type: array
          uniqueItems: true
          items:
            type: string
            enum: [platform_admin, department_admin, document_admin, auditor]
    UserRoleUpdateResponse:
      type: object
      additionalProperties: false
      required: [user, traceId]
      properties:
        user: { $ref: '#/components/schemas/UserDirectoryEntry' }
        traceId: { type: string, format: uuid }
    DepartmentPolicy:
      type: object
      additionalProperties: false
      required: [department, allowedSensitivities, userCount, documentCount, managed, updatedAt]
      properties:
        department: { type: string, minLength: 1, maxLength: 128 }
        allowedSensitivities:
          type: array
          minItems: 1
          maxItems: 3
          items: { $ref: '#/components/schemas/Sensitivity' }
        userCount: { type: integer, minimum: 0 }
        documentCount: { type: integer, minimum: 0 }
        managed: { type: boolean }
        updatedAt: { type: [string, 'null'], format: date-time }
    DepartmentPolicyListResponse:
      type: object
      additionalProperties: false
      required: [departments, scope]
      properties:
        departments:
          type: array
          items: { $ref: '#/components/schemas/DepartmentPolicy' }
        scope: { type: string, enum: [tenant, department] }
    DepartmentPolicyUpdateRequest:
      type: object
      additionalProperties: false
      required: [allowedSensitivities]
      properties:
        allowedSensitivities:
          type: array
          minItems: 1
          maxItems: 3
          items: { $ref: '#/components/schemas/Sensitivity' }
    DepartmentPolicyUpdateResponse:
      type: object
      additionalProperties: false
      required: [department, traceId]
      properties:
        department: { $ref: '#/components/schemas/DepartmentPolicy' }
        traceId: { type: string, format: uuid }
    ConversationSummary:
      type: object
      additionalProperties: false
      required: [id, title, messageCount, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        title: { type: string, minLength: 1, maxLength: 200 }
        messageCount: { type: integer, minimum: 0 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    ConversationListResponse:
      type: object
      additionalProperties: false
      required: [conversations, total, offset, limit]
      properties:
        conversations:
          type: array
          maxItems: 100
          items: { $ref: '#/components/schemas/ConversationSummary' }
        total: { type: integer, minimum: 0 }
        offset: { type: integer, minimum: 0 }
        limit: { type: integer, minimum: 1, maximum: 100 }
    ConversationTurn:
      type: object
      additionalProperties: false
      required: [id, question, answer, noAnswer, reason, traceId, sourceCount, createdAt]
      properties:
        id: { type: string, format: uuid }
        question: { type: string, minLength: 1, maxLength: 2000 }
        answer: { type: string, minLength: 1 }
        noAnswer: { type: boolean }
        reason:
          { type: [string, 'null'], enum: [insufficient_relevance, authorization_changed, null] }
        traceId: { type: string, format: uuid }
        sourceCount: { type: integer, minimum: 0 }
        createdAt: { type: string, format: date-time }
    ConversationDetail:
      allOf:
        - { $ref: '#/components/schemas/ConversationSummary' }
        - type: object
          required: [turns]
          properties:
            turns:
              type: array
              maxItems: 500
              items: { $ref: '#/components/schemas/ConversationTurn' }
    ConversationDeleteResponse:
      type: object
      additionalProperties: false
      required: [conversationId, deleted]
      properties:
        conversationId: { type: string, format: uuid }
        deleted: { const: true }
...
    UploadAccepted:
      type: object
      required: [documentId, jobId, status, traceId]
      properties:
        documentId: { type: string, format: uuid }
        jobId: { type: string, format: uuid }
        status: { const: queued }
        traceId: { type: string, format: uuid }
    ReindexAccepted:
      type: object
      additionalProperties: false
      required: [documentId, documentVersion, jobId, status, traceId]
      properties:
        documentId: { type: string, format: uuid }
        documentVersion: { type: integer, minimum: 2 }
        jobId: { type: string, format: uuid }
        status: { const: queued }
        traceId: { type: string, format: uuid }
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, traceId]
          properties:
            code: { type: string }
            message: { type: string }
            traceId: { type: string, format: uuid }
    DocumentStatus:
      type: string
      enum: [uploaded, processing, prepared, active, policy_blocked, failed, deleting, deleted]
    DocumentFormat:
      type: string
      enum: [txt, md, docx, xlsx, dxf, dwg]
    DocumentListItem:
      type: object
      additionalProperties: false
      required:
        [
          id,
          sourceName,
          mimeType,
          department,
          sensitivity,
          ownerId,
          activeVersion,
          status,
          latestJob,
          createdAt,
          updatedAt,
        ]
      properties:
        id: { type: string, format: uuid }
        sourceName: { type: string }
        mimeType: { type: string }
        department: { type: string }
        sensitivity: { $ref: '#/components/schemas/Sensitivity' }
        ownerId: { type: string }
        activeVersion: { type: [integer, 'null'], minimum: 1 }
        status: { $ref: '#/components/schemas/DocumentStatus' }
        latestJob:
          oneOf:
            - { $ref: '#/components/schemas/LatestIngestionJob' }
            - { type: 'null' }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    LatestIngestionJob:
      type: object
      additionalProperties: false
      required: [id, status, step, attempts, retryable, errorCode, updatedAt]
      properties:
        id: { type: string, format: uuid }
        status: { $ref: '#/components/schemas/IngestionStatus' }
        step: { type: string }
        attempts: { type: integer, minimum: 0 }
        retryable: { type: boolean }
        errorCode: { type: [string, 'null'] }
        updatedAt: { type: string, format: date-time }
    DocumentListResponse:
      type: object
      additionalProperties: false
      required: [items, page, pageSize, total]
      properties:
        items:
          { type: array, maxItems: 100, items: { $ref: '#/components/schemas/DocumentListItem' } }
        page: { type: integer, minimum: 1 }
        pageSize: { type: integer, minimum: 1, maximum: 100 }
        total: { type: integer, minimum: 0 }
    DocumentUploadOptions:
      type: object
      additionalProperties: false
      required:
        [
          maxUploadBytes,
          acceptedExtensions,
          department,
          allowedSensitivities,
          defaultSensitivity,
          dwgConversionEnabled,
        ]
      properties:
        maxUploadBytes: { type: integer, minimum: 1 }
        acceptedExtensions:
          { type: array, minItems: 1, items: { $ref: '#/components/schemas/DocumentFormat' } }
        department: { type: string }
        allowedSensitivities:
          { type: array, minItems: 1, items: { $ref: '#/components/schemas/Sensitivity' } }
        defaultSensitivity: { $ref: '#/components/schemas/Sensitivity' }
        dwgConversionEnabled: { type: boolean }
    DocumentVersion:
      type: object
      additionalProperties: false
      required:
        [
          version,
          status,
          parser,
          parserVersion,
          warnings,
          chunkCount,
          embeddingFingerprint,
          indexedAt,
          activatedAt,
          supersededAt,
          createdAt,
        ]
      properties:
        version: { type: integer, minimum: 1 }
        status:
          {
            type: string,
            enum: [processing, prepared, policy_blocked, active, superseded, failed, deleted],
          }
        parser: { type: [string, 'null'] }
        parserVersion: { type: [string, 'null'] }
        warnings: { type: array, items: { type: string } }
        chunkCount: { type: integer, minimum: 0 }
        embeddingFingerprint: { type: [string, 'null'], pattern: '^[0-9a-f]{64}$' }
        indexedAt: { type: [string, 'null'], format: date-time }
        activatedAt: { type: [string, 'null'], format: date-time }
        supersededAt: { type: [string, 'null'], format: date-time }
        createdAt: { type: string, format: date-time }
    DocumentDetail:
      type: object
      additionalProperties: false
      required:
        [
          id,
          sourceName,
          mimeType,
          department,
          sensitivity,
          ownerId,
          activeVersion,
          status,
          versions,
          createdAt,
          updatedAt,
        ]
      properties:
        id: { type: string, format: uuid }
        sourceName: { type: string }
        mimeType: { type: string }
        department: { type: string }
        sensitivity: { $ref: '#/components/schemas/Sensitivity' }
        ownerId: { type: string }
        activeVersion: { type: [integer, 'null'], minimum: 1 }
        status: { $ref: '#/components/schemas/DocumentStatus' }
        versions: { type: array, items: { $ref: '#/components/schemas/DocumentVersion' } }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    DocumentDeleteResponse:
      type: object
      additionalProperties: false
      required: [documentId, deleted]
      properties:
        documentId: { type: string, format: uuid }
        deleted: { const: true }
    IngestionStatus:
      type: string
      enum:
        [
          queued,
          converting,
          parsing,
          chunking,
          policy_check,
          embedding,
          indexing,
          policy_blocked,
          completed,
          failed,
          deleted,
        ]
    IngestionJobSummary:
      type: object
      required:
        [
          id,
          documentId,
          version,
          kind,
          status,
          step,
          checkpoint,
          attempts,
          traceId,
          errorCode,
          errorCategory,
          retryable,
          startedAt,
          completedAt,
          updatedAt,
        ]
      properties:
        id: { type: string, format: uuid }
        documentId: { type: string, format: uuid }
        version: { type: integer, minimum: 1 }
        kind: { type: string, enum: [ingestion, reindex, index_migration] }
        status: { $ref: '#/components/schemas/IngestionStatus' }
        step: { type: string }
        checkpoint: { type: string }
        attempts: { type: integer, minimum: 0 }
        traceId: { type: string, format: uuid }
        errorCode: { type: [string, 'null'] }
        errorCategory: { type: [string, 'null'] }
        retryable: { type: boolean }
        startedAt: { type: [string, 'null'], format: date-time }
        completedAt: { type: [string, 'null'], format: date-time }
        updatedAt: { type: string, format: date-time }
    IngestionJob:
      allOf:
        - { $ref: '#/components/schemas/IngestionJobSummary' }
        - type: object
          required: [sourceName, mimeType, parserVersion, embeddingFingerprint, warnings, createdAt]
          properties:
            sourceName: { type: string }
            mimeType: { type: string }
            parserVersion: { type: [string, 'null'] }
            embeddingFingerprint: { type: [string, 'null'], pattern: '^[0-9a-f]{64}$' }
            warnings: { type: array, items: { type: string } }
            createdAt: { type: string, format: date-time }
    IngestionJobListResponse:
      type: object
      additionalProperties: false
      required: [items, page, pageSize, total]
      properties:
        items: { type: array, maxItems: 100, items: { $ref: '#/components/schemas/IngestionJob' } }
        page: { type: integer, minimum: 1 }
        pageSize: { type: integer, minimum: 1, maximum: 100 }
        total: { type: integer, minimum: 0 }
    IngestionRetryAccepted:
      type: object
      additionalProperties: false
      required: [jobId, status, traceId]
      properties:
        jobId: { type: string, format: uuid }
        status: { const: queued }
        traceId: { type: string, format: uuid }
  responses:
    Unauthorized:
      description: Missing, invalid or expired bearer token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    Forbidden:
      description: Authenticated identity lacks the required capability
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    NotFound:
      description: Resource does not exist in the authenticated tenant
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }